Add privatePath for assembly probing
